+++ src/main/resources/db/migration/20180118154515853__added_locally_fulfilled_flag_to_supported_programs.sql
@@ -1,0 +1,2 @@
+ALTER TABLE referencedata.supported_programs
+ADD COLUMN locallyFulfilled boolean NOT NULL DEFAULT FALSE;